made scoreboard test
