+++ lib/scoreboard.js
@@ -21,11 +21,12 @@ Scoreboard.prototype.savePoints = function(points){
 };
 
 Scoreboard.prototype.loadScoreboard = function(){
+  var scoreboard = this
   scoreboardRecords = [];
   $('#scoreboard-records').empty();
 
   fireDb.ref('highscore/').once('value').then(function(scores){
-    var allScores = sortScores(scores.val().highscores);
+    var allScores = scoreboard.sortScores(scores.val().highscores);
 
     for (var i = 0; i < 5; i++) {
       if (allScores[i] !== undefined) {
@@ -39,7 +40,7 @@ Scoreboard.prototype.loadScoreboard = function(){
   });
 };
 
-function sortScores(scores){
+Scoreboard.prototype.sortScores = function(scores){
   var sortedScores = scores.sort(function(a, b){return b.points - a.points})
   return sortedScores
 };
