+++ test/index.js
@@ -5,3 +5,4 @@ require('./trash-test');
 require('./game-test');
 require('./helpers-test');
 require('./background-test');
+require('./scoreboard-test');
